Imaginary Projection

Imaginary Projection

Amoebas.imaginary_projection — Function.

imaginary_projection(f; alg=Greedy(), options...)

Compute the imaginary_projection of f with the given algorithm alg which can be

Greedy()

Simple()

but you probably only want to use Greedy().

Examples

@polyvar x y

# This uses the `Greedy()` algorithm
amoeba(x^2 + y^2 + 1)

# Use the `Greedy()` algorithm with a custom grid.
grid = Grid2D(xlims=(-5, 5), ylims=(-4, 4), res=(500, 400))
amoeba(x^2 + y^2 + 1, alg=Greedy(), grid=grid)

# Use the `Greedy()` algorithm  with the default domain but a higher resolution
amoeba(x^2 + y^2 + 1, alg=Greedy(), resolution=800)

These algorithms are all approximations of the imaginary projection 𝐼(𝑓) based on a grid. This basically applies the membership test for different grid points. The grid can be passed explicitly, otherwise it will be computed based on a heuristic.

resolution=1000: The resolution of the grid if not passed explicitly. If the Greedy() algorithm is

used then a higher resolution can improve the image quality substantially.

grid: If passed explicitly this grid is used, otherwise it will be computed based on a heuristic.

membership_options=[MembershipTestOptions()](@ref): The options for the membership test.

npasses=1: The Greedy() algorithm can make multiple passes to improve the quality.